# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a table top game.",
        "title": "Table Top API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/special_effects": {
            "get": {
                "description": "List all special effects",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SpecialEffects"
                ],
                "summary": "List all special effects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SpecialEffect"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new special effect",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SpecialEffects"
                ],
                "summary": "Create a new special effect",
                "parameters": [
                    {
                        "description": "Special Effect",
                        "name": "specialEffect",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/special_effects.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SpecialEffect"
                        }
                    }
                }
            }
        },
        "/special_effects/{id}": {
            "get": {
                "description": "Get a special effect by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SpecialEffects"
                ],
                "summary": "Get a special effect by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Special Effect ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SpecialEffect"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a special effect",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SpecialEffects"
                ],
                "summary": "Update a special effect",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Special Effect ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Special Effect",
                        "name": "specialEffect",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/special_effects.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SpecialEffect"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a special effect",
                "tags": [
                    "SpecialEffects"
                ],
                "summary": "Delete a special effect",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Special Effect ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/weapon_combos": {
            "get": {
                "description": "List all weapon combos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WeaponCombos"
                ],
                "summary": "List all weapon combos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.WeaponCombo"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new weapon combo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WeaponCombos"
                ],
                "summary": "Create a new weapon combo",
                "parameters": [
                    {
                        "description": "WeaponCombo",
                        "name": "weaponCombo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/weapon_combos.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.WeaponCombo"
                        }
                    }
                }
            }
        },
        "/weapon_combos/{id}": {
            "get": {
                "description": "Get a weapon combo by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WeaponCombos"
                ],
                "summary": "Get a weapon combo by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "WeaponCombo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.WeaponCombo"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a weapon combo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WeaponCombos"
                ],
                "summary": "Update a weapon combo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "WeaponCombo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "WeaponCombo",
                        "name": "weaponCombo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/weapon_combos.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.WeaponCombo"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a weapon combo by ID",
                "tags": [
                    "WeaponCombos"
                ],
                "summary": "Delete a weapon combo by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "WeaponCombo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.SpecialEffect": {
            "type": "object",
            "properties": {
                "combo_card_id": {
                    "type": "string"
                },
                "damage": {
                    "type": "integer"
                },
                "damage_type": {
                    "$ref": "#/definitions/models.SpecialEffectDamageType"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.SpecialEffectType"
                }
            }
        },
        "models.SpecialEffectDamageType": {
            "type": "integer",
            "enum": [
                0
            ],
            "x-enum-varnames": [
                "EffectOnly"
            ]
        },
        "models.SpecialEffectType": {
            "type": "integer",
            "enum": [
                0
            ],
            "x-enum-varnames": [
                "Variable"
            ]
        },
...
        "models.WeaponCombo": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/models.WeaponComboType"
                },
                "weapon_id": {
                    "type": "string"
                }
            }
        },
        "models.WeaponComboType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "FlexSlot",
                "CombatSkillSlot",
                "AbilitySlot"
            ]
        },
        "special_effects.CreateRequest": {
            "type": "object",
            "properties": {
                "combo_card_id": {
                    "type": "string"
                },
                "damage": {
                    "type": "integer"
                },
                "damage_type": {
                    "$ref": "#/definitions/models.SpecialEffectDamageType"
                },
                "description": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.SpecialEffectType"
                }
            }
        },
        "special_effects.UpdateRequest": {
            "type": "object",
            "properties": {
                "combo_card_id": {
                    "type": "string"
                },
                "damage": {
                    "type": "integer"
                },
                "damage_type": {
                    "$ref": "#/definitions/models.SpecialEffectDamageType"
                },
                "description": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.SpecialEffectType"
                }
            }
        },
        "weapon_combos.CreateRequest": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "order": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/models.WeaponComboType"
                },
                "weapon_id": {
                    "type": "string"
                }
            }
        },
        "weapon_combos.UpdateRequest": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "order": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/models.WeaponComboType"
                },
                "weapon_id": {
                    "type": "string"
                }
            }
        },
...
    }
}